refactor(dumpsxiso): extract XML generation logic to xmlwriter
