+++ src/dumpsxiso/main.cpp
@@ -1,7 +1,6 @@
 #include "platform.h"
-#include "xml.h"
+#include "xmlwriter.h"
 #include "cue.h"
-#include <map>
 
 #ifndef MKPSXISO_NO_LIBFLAC
 #include "FLAC/stream_encoder.h"
@@ -58,6 +57,7 @@ namespace global
 	CueFile cueFile;
 	bool ps2 = false;
 	bool xa_edc = true;
+	std::string licenseFile;
 	std::optional<bool> cdvd_style;
 }
 
@@ -106,21 +106,6 @@ void PrintDate(const char* label, const cd::ISO_LONG_DATESTAMP& date)
 	{
 		printf("%s%s\n", label, LongDateToString(date).c_str());
 	}
-}
-
-template<size_t N>
-static std::string_view CleanDescElement(const char (&id)[N])
-{
-	std::string_view result;
-
-	const std::string_view view(id, N);
-	const size_t last_char = view.find_last_not_of(' ');
-	if (last_char != std::string_view::npos)
-	{
-		result = view.substr(0, last_char+1);
-	}
-
-    return result;
 }
 
 void prepareRIFFHeader(cd::RIFF_HEADER* header, int dataSize) {
@@ -195,7 +180,8 @@ void SaveLicense(const cd::ISO_LICENSE& license) {
 		printf("\n  Creating license data...");
 	}
 
-	FILE* outFile = OpenFile(param::outPath / "license_data.dat", "wb");
+	global::licenseFile = "license_data.dat";
+	FILE* outFile = OpenFile(param::outPath / global::licenseFile, "wb");
 
 	if (outFile == NULL)
 	{
@@ -317,119 +303,6 @@ writeFLACFile_cleanup:
 	}
 }
 #endif
-
-// XML attribute stuff
-struct EntryAttributeCounters
-{
-	std::map<int, unsigned int> GMTOffs;
-	std::map<int, unsigned int> HFLAG;
-	std::map<int, unsigned int> XAAttrib;
-	std::map<int, unsigned int> XAPerm;
-	std::map<int, unsigned int> GID;
-	std::map<int, unsigned int> UID;
-};
-
-static void WriteOptionalXMLAttribs(tinyxml2::XMLElement* element, const cd::IsoDirEntries::Entry& entry, EntryType type, EntryAttributeCounters& attributeCounters)
-{
-	element->SetAttribute(xml::attrib::GMT_OFFSET, entry.entry.entryDate.GMToffs);
-	++attributeCounters.GMTOffs[entry.entry.entryDate.GMToffs];
-
-	// xa_attrib only makes sense on XA files
-	if (type == EntryType::EntryXA)
-	{
-		const auto XAAtrib = (entry.extData.attributes & cdxa::XA_ATTRIBUTES_MASK) >> 8;
-		element->SetAttribute(xml::attrib::XA_ATTRIBUTES, XAAtrib);
-		++attributeCounters.XAAttrib[XAAtrib];
-	}
-	const auto XAPerm = entry.extData.attributes & cdxa::XA_PERMISSIONS_MASK;
-	element->SetAttribute(xml::attrib::XA_PERMISSIONS, XAPerm);
-	++attributeCounters.XAPerm[XAPerm];
-
-	element->SetAttribute(xml::attrib::XA_GID, entry.extData.ownergroupid);
-	element->SetAttribute(xml::attrib::XA_UID, entry.extData.owneruserid);
-	++attributeCounters.GID[entry.extData.ownergroupid];
-	++attributeCounters.UID[entry.extData.owneruserid];
-
-	const auto HFLAG = entry.entry.flags & 0x20 ? entry.entry.flags & 0x03 : entry.entry.flags & 0x01; // 5th bit simulates obfuscation
-	element->SetAttribute(xml::attrib::HIDDEN_FLAG, HFLAG);
-	++attributeCounters.HFLAG[HFLAG];
-
-	if (entry.order.has_value())
-	{
-		element->SetAttribute(xml::attrib::ORDER, *entry.order);
-	}
-	if (param::lba)
-	{
-		element->SetAttribute(xml::attrib::OFFSET, entry.entry.entryOffs.lsb);
-	}
-}
-
-static EntryAttributes EstablishXMLAttributeDefaults(tinyxml2::XMLElement* defaultAttributesElement, const EntryAttributeCounters& attributeCounters)
-{
-	// First establish "defaults" - that is, the most commonly occurring attributes
-	auto findMaxElement = [](const auto& map)
-	{
-		if (!map.empty()) {
-			return std::max_element(map.begin(), map.end(), [](const auto& left, const auto& right) { return left.second < right.second; })->first;
-		}
-		return 0;
-	};
-
-	EntryAttributes defaultAttributes;
-	defaultAttributes.GMTOffs = static_cast<signed char>(findMaxElement(attributeCounters.GMTOffs));
-	defaultAttributes.HFLAG = static_cast<unsigned char>(findMaxElement(attributeCounters.HFLAG));
-	defaultAttributes.XAAttrib = static_cast<unsigned char>(findMaxElement(attributeCounters.XAAttrib));
-	defaultAttributes.XAPerm = static_cast<unsigned short>(findMaxElement(attributeCounters.XAPerm));
-	defaultAttributes.GID = static_cast<unsigned short>(findMaxElement(attributeCounters.GID));
-	defaultAttributes.UID = static_cast<unsigned short>(findMaxElement(attributeCounters.UID));
-
-	// Write them out to the XML
-	defaultAttributesElement->SetAttribute(xml::attrib::GMT_OFFSET, defaultAttributes.GMTOffs);
-	defaultAttributesElement->SetAttribute(xml::attrib::XA_ATTRIBUTES, defaultAttributes.XAAttrib);
-	defaultAttributesElement->SetAttribute(xml::attrib::XA_PERMISSIONS, defaultAttributes.XAPerm);
-	defaultAttributesElement->SetAttribute(xml::attrib::XA_GID, defaultAttributes.GID);
-	defaultAttributesElement->SetAttribute(xml::attrib::XA_UID, defaultAttributes.UID);
-	if (defaultAttributes.HFLAG) // Set only if not zero
-	{
-		defaultAttributesElement->SetAttribute(xml::attrib::HIDDEN_FLAG, defaultAttributes.HFLAG);
-	}
-
-	return defaultAttributes;
-}
-
-static void SimplifyDefaultXMLAttributes(tinyxml2::XMLElement* element, const EntryAttributes& defaults)
-{
-	// DeleteAttribute can be safely called even if that attribute doesn't exist, so treating failure and default values
-	// as equal simplifies logic
-	auto deleteAttribute = [element](const char* name, auto defaultValue)
-	{
-		bool deleteAttribute = false;
-		if constexpr (std::is_unsigned_v<decltype(defaultValue)>)
-		{
-			deleteAttribute = element->UnsignedAttribute(name, defaultValue) == defaultValue;
-		}
-		else
-		{
-			deleteAttribute = element->IntAttribute(name, defaultValue) == defaultValue;
-		}
-		if (deleteAttribute)
-		{
-			element->DeleteAttribute(name);
-		}
-	};
-
-	deleteAttribute(xml::attrib::GMT_OFFSET, defaults.GMTOffs);
-	deleteAttribute(xml::attrib::HIDDEN_FLAG, defaults.HFLAG);
-	deleteAttribute(xml::attrib::XA_ATTRIBUTES, defaults.XAAttrib);
-	deleteAttribute(xml::attrib::XA_PERMISSIONS, defaults.XAPerm);
-	deleteAttribute(xml::attrib::XA_GID, defaults.GID);
-	deleteAttribute(xml::attrib::XA_UID, defaults.UID);
-
-	for (tinyxml2::XMLElement* child = element->FirstChildElement(); child != nullptr; child = child->NextSiblingElement())
-	{
-		SimplifyDefaultXMLAttributes(child, defaults);
-	}
-}
 
 std::unique_ptr<cd::IsoDirEntries> ParseSubdirectory(cd::IsoReader& reader, ListView<cd::IsoDirEntries::Entry> view, int offs, int sectors,
 	const fs::path& path)
@@ -889,186 +762,6 @@ void ExtractFiles(cd::IsoReader& reader, const std::list<cd::IsoDirEntries::Entr
 	}
 }
 
-tinyxml2::XMLElement* WriteXMLEntry(const cd::IsoDirEntries::Entry& entry, tinyxml2::XMLElement* dirElement, fs::path* currentVirtualPath,
-	const fs::path& sourcePath, EntryAttributeCounters& attributeCounters)
-{
-	tinyxml2::XMLElement* newelement;
-
-	if (entry.type == EntryType::EntryDir)
-	{
-		if (!entry.identifier.empty())
-		{
-			newelement = dirElement->InsertNewChildElement("dir");
-			newelement->SetAttribute(xml::attrib::ENTRY_NAME, entry.identifier.c_str());
-			if (param::lba)
-			{
-				const fs::path outputPath = sourcePath / entry.virtualPath / CleanIdentifier(entry.identifier);
-				newelement->SetAttribute(xml::attrib::ENTRY_SOURCE, outputPath.generic_string().c_str());
-			}
-			newelement->SetAttribute(xml::attrib::ENTRY_DATE, DateToString(entry.entry.entryDate, false).c_str());
-		}
-		else
-		{
-			// Root directory
-			newelement = dirElement->InsertNewChildElement(xml::elem::DIRECTORY_TREE);
-			if (!param::lba)
-			{
-				newelement->SetAttribute(xml::attrib::ENTRY_SOURCE, sourcePath.generic_string().c_str());
-			}
-		}
-
-		dirElement = newelement;
-		if (currentVirtualPath != nullptr)
-		{
-			*currentVirtualPath /= entry.identifier;
-		}
-    }
-	else
-	{
-        newelement = dirElement->InsertNewChildElement("file");
-		newelement->SetAttribute(xml::attrib::ENTRY_NAME, CleanIdentifier(entry.identifier).c_str());
-		if(entry.type != EntryType::EntryDA)
-		{
-			if (param::lba)
-			{
-				const fs::path outputPath = sourcePath / entry.virtualPath / CleanIdentifier(entry.identifier);
-				newelement->SetAttribute(xml::attrib::ENTRY_SOURCE, outputPath.generic_string().c_str());
-			}
-			newelement->SetAttribute(xml::attrib::ENTRY_TYPE, entry.type == EntryType::EntryFile ? "data" : "mixed");
-		}
-		else
-		{
-			newelement->SetAttribute(xml::attrib::TRACK_ID, entry.trackid.c_str());
-			newelement->SetAttribute(xml::attrib::ENTRY_TYPE, "da");
-		}
-		newelement->SetAttribute(xml::attrib::ENTRY_DATE, DateToString(entry.entry.entryDate, false).c_str());
-	}
-	WriteOptionalXMLAttribs(newelement, entry, entry.type, attributeCounters);
-	return dirElement;
-}
-
-void WriteXMLGap(const unsigned int numSectors, tinyxml2::XMLElement* dirElement, const unsigned int startSector, cd::IsoReader &reader)
-{
-	if (numSectors < 1) {
-		return;
-	}
-	cd::SECTOR_M2F1 sector;
-	reader.SeekToSector(startSector);
-	reader.ReadBytesXA(sector.subHead, XA_DATA_SIZE);
-	tinyxml2::XMLElement* newelement = dirElement->InsertNewChildElement("dummy");
-	newelement->SetAttribute(xml::attrib::NUM_DUMMY_SECTORS, numSectors);
-	newelement->SetAttribute(xml::attrib::ENTRY_TYPE, sector.subHead[2]);
-	if (param::lba)
-	{
-		newelement->SetAttribute(xml::attrib::OFFSET, startSector);
-	}
-}
-
-void WriteXMLPostGap(const unsigned int& postGap, tinyxml2::XMLElement* dirTree, unsigned int& currentLBA, cd::IsoReader& reader)
-{
-	if (!postGap)
-		return;
-
-	// There are some CD-DA games that have a non-zero adrress ECC calculation in the last postgap sector. So, we are checking it.
-	// Idk if this behavior could happen in other sectors, but apparently it's only related to CD-DA games postgaps (maybe a bug).
-	cd::SECTOR_M2F1 sector;
-	reader.SeekToSector(currentLBA + postGap - 1);
-	reader.ReadBytesXA(sector.subHead, XA_DATA_SIZE);
-	if (memcmp(sector.ecc, "\0\0\0\0", sizeof(sector.edc)))
-	{
-		WriteXMLGap(postGap - 1, dirTree, currentLBA, reader);
-		WriteXMLGap(1, dirTree, currentLBA + postGap - 1, reader);
-		dirTree->LastChildElement()->SetAttribute(xml::attrib::ECC_ADDRES, true);
-	}
-	else
-	{
-		WriteXMLGap(postGap, dirTree, currentLBA, reader);
-	}
-	currentLBA += postGap;
-}
-
-void WriteXMLByLBA(const std::list<cd::IsoDirEntries::Entry>& files, tinyxml2::XMLElement* dirElement, const fs::path& sourcePath, unsigned int& expectedLBA,
-	EntryAttributeCounters& attributeCounters, cd::IsoReader& reader, const unsigned int& postGap)
-{
-	fs::path currentVirtualPath; // Used to find out whether to traverse 'dir' up or down the chain
-	bool writedPostGap = false;
-	for (const auto& entry : files)
-	{
-		// if this is a DA file we are at the end of filesystem
-		if (entry.type != EntryType::EntryDA)
-		{
-			// only check for gaps, update LBA if it's inside the iso filesystem
-			if (entry.entry.entryOffs.lsb > expectedLBA)
-			{
-				WriteXMLGap(entry.entry.entryOffs.lsb - expectedLBA, dirElement, expectedLBA, reader);
-			}
-			expectedLBA = entry.entry.entryOffs.lsb + GetSizeInSectors(entry.entry.entrySize.lsb);
-		}
-		else if (entry.trackid.empty())
-		{
-			continue; // Skip if it's an unreferenced DA file
-		}
-		else if (!writedPostGap)
-		{
-			WriteXMLPostGap(postGap, dirElement, expectedLBA, reader);
-			writedPostGap = true;
-		}
-
-		// Work out the relative position between the current directory and the element to create
-		const fs::path relative = entry.virtualPath.lexically_relative(currentVirtualPath);
-		for (const fs::path& part : relative)
-		{
-			if (part == "..")
-			{
-				// Go up in XML
-				dirElement = dirElement->Parent()->ToElement();
-				currentVirtualPath = currentVirtualPath.parent_path();
-				continue;
-			}
-			if (part == ".")
-			{
-				// Do nothing
-				continue;
-			}
-
-			// "Enter" the directory
-			dirElement = dirElement->InsertNewChildElement("dir");
-			dirElement->SetAttribute(xml::attrib::ENTRY_NAME, part.generic_string().c_str());
-
-			currentVirtualPath /= part;
-		}
-
-		dirElement = WriteXMLEntry(entry, dirElement, &currentVirtualPath, sourcePath, attributeCounters);
-	}
-
-	if (!writedPostGap)
-	{
-		WriteXMLPostGap(postGap, dirElement, expectedLBA, reader);
-	}
-}
-
-void WriteXMLByDirectories(const cd::IsoDirEntries* directory, tinyxml2::XMLElement* dirElement, const fs::path& sourcePath, unsigned int& expectedLBA,
-	EntryAttributeCounters& attributeCounters)
-{
-	for (const auto& e : directory->dirEntryList.GetView())
-	{
-		const auto& entry = e.get();
-
-		if (entry.type != EntryType::EntryDA)
-		{
-			// Update the LBA to the max encountered value
-			expectedLBA = std::max(expectedLBA, entry.entry.entryOffs.lsb + GetSizeInSectors(entry.entry.entrySize.lsb));
-		}
-
-		tinyxml2::XMLElement* child = WriteXMLEntry(entry, dirElement, nullptr, sourcePath, attributeCounters);
-		// Recursively write children if there are any
-		if (const cd::IsoDirEntries* subdir = entry.subdir.get(); subdir != nullptr)
-		{
-			WriteXMLByDirectories(subdir, child, sourcePath, expectedLBA, attributeCounters);
-		}
-	}
-}
-
 void ParseISO(cd::IsoReader& reader) {
 
     cd::ISO_DESCRIPTOR descriptor;
@@ -1208,133 +901,27 @@ void ParseISO(cd::IsoReader& reader) {
 		{
 			printf("  Creating XML document...");
 		}
-		if (FILE* file = OpenFile(param::xmlFile, "wb"); file != nullptr)
+
+		const unsigned currentLBA = xml::WriteXML(descriptor, rootDir, DAfiles, postGap);
+
+		if (!param::QuietMode)
 		{
-			if (!param::QuietMode)
+			printf(" Ok.\n\n");
+		}
+
+		// Check if there is still an EoF gap
+		if (!param::noWarns && (postGap > 150 || currentLBA < totalLenLBA))
+		{
+			printf( "WARNING: There is still a gap of %u sectors at the end of file system.\n"
+					"\t This could mean that there are missing files or tracks.\n", postGap > 150 ? postGap : totalLenLBA - currentLBA);
+			if (global::cueFile.tracks.empty())
 			{
-				printf(" Ok.\n\n");
-			}
-			tinyxml2::XMLDocument xmldoc;
-
-			tinyxml2::XMLElement *baseElement = static_cast<tinyxml2::XMLElement*>(xmldoc.InsertFirstChild(xmldoc.NewElement(xml::elem::ISO_PROJECT)));
-			baseElement->SetAttribute(xml::attrib::IMAGE_NAME, "mkpsxiso.bin");
-			baseElement->SetAttribute(xml::attrib::CUE_SHEET, "mkpsxiso.cue");
-
-			tinyxml2::XMLElement *trackElement = baseElement->InsertNewChildElement(xml::elem::TRACK);
-			trackElement->SetAttribute(xml::attrib::TRACK_TYPE, "data");
-			trackElement->SetAttribute(xml::attrib::XA_EDC, global::xa_edc);
-			trackElement->SetAttribute(xml::attrib::CDVD_STYLE, *global::cdvd_style);
-			if (global::ps2)
-			{
-				trackElement->SetAttribute(xml::attrib::PS2, global::ps2);
-			}
-
-			{
-				tinyxml2::XMLElement *newElement = trackElement->InsertNewChildElement(xml::elem::IDENTIFIERS);
-				auto setAttributeIfNotEmpty = [newElement](const char* name, std::string_view value)
-				{
-					if (!value.empty())
-					{
-						newElement->SetAttribute(name, std::string(value).c_str());
-					}
-				};
-
-				setAttributeIfNotEmpty(xml::attrib::SYSTEM_ID, CleanDescElement(descriptor.systemID));
-				setAttributeIfNotEmpty(xml::attrib::APPLICATION, CleanDescElement(descriptor.applicationIdentifier));
-				setAttributeIfNotEmpty(xml::attrib::VOLUME_ID, CleanDescElement(descriptor.volumeID));
-				setAttributeIfNotEmpty(xml::attrib::VOLUME_SET, CleanDescElement(descriptor.volumeSetIdentifier));
-				setAttributeIfNotEmpty(xml::attrib::PUBLISHER, CleanDescElement(descriptor.publisherIdentifier));
-				setAttributeIfNotEmpty(xml::attrib::DATA_PREPARER, CleanDescElement(descriptor.dataPreparerIdentifier));
-				setAttributeIfNotEmpty(xml::attrib::COPYRIGHT, CleanDescElement(descriptor.copyrightFileIdentifier));
-				newElement->SetAttribute(xml::attrib::CREATION_DATE, LongDateToString(descriptor.volumeCreateDate).c_str());
-				if (auto ZERO_DATE = GetUnspecifiedLongDate(); memcmp(&descriptor.volumeModifyDate, &ZERO_DATE, sizeof(descriptor.volumeModifyDate)) != 0)
-				{
-					// Set only if not zero
-					newElement->SetAttribute(xml::attrib::MODIFICATION_DATE, LongDateToString(descriptor.volumeModifyDate).c_str());
-				}
-			}
-
-			const fs::path xmlPath = param::xmlFile.parent_path();
-			const fs::path sourcePath = xmlPath.is_absolute() ? fs::absolute(param::outPath) : param::outPath.lexically_proximate(xmlPath);
-
-			// Add license element to the xml
-			{
-				tinyxml2::XMLElement *newElement = trackElement->InsertNewChildElement(xml::elem::LICENSE);
-				newElement->SetAttribute(xml::attrib::LICENSE_FILE,	(sourcePath / "license_data.dat").generic_string().c_str());
-			}
-
-			// Create <default_attributes> now so it lands before the directory tree
-			tinyxml2::XMLElement* defaultAttributesElement = trackElement->InsertNewChildElement(xml::elem::DEFAULT_ATTRIBUTES);
-
-			EntryAttributeCounters attributeCounters;
-			unsigned currentLBA = descriptor.rootDirRecord.entryOffs.lsb;
-			if (param::outputSortedByDir)
-			{
-				WriteXMLByDirectories(rootDir.get(), trackElement, sourcePath, currentLBA, attributeCounters);		
-				WriteXMLPostGap(postGap, trackElement->LastChildElement(), currentLBA, reader);
+				printf("\t Try using a .cue file instead of an ISO image.\n");
 			}
 			else
 			{
-				WriteXMLByLBA(entries, trackElement, sourcePath, currentLBA, attributeCounters, reader, postGap);
+				printf("\t Try using the -pt or/and -f command, helps with obfuscated file systems.\n");
 			}
-
-			tinyxml2::XMLElement *dirtree = trackElement->FirstChildElement(xml::elem::DIRECTORY_TREE);
-			SimplifyDefaultXMLAttributes(dirtree, EstablishXMLAttributeDefaults(defaultAttributesElement, attributeCounters));
-
-			// Write CD-DA tracks
-			tinyxml2::XMLNode *modifyProject = trackElement->Parent();
-			tinyxml2::XMLElement *addAfter = trackElement;
-			for(const auto& dafile : DAfiles)
-			{
-				// SYSTEM DESCRIPTION CD-ROM XA Ch.II 2.3, pause should be always >= 150 sectors.
-				unsigned pregap_sectors = 150;
-				dafile->virtualPath = GetEncodedDAPath(sourcePath / dafile->virtualPath / CleanIdentifier(dafile->identifier));
-				if(dafile->entry.entryOffs.lsb != currentLBA)
-				{
-					pregap_sectors = dafile->entry.entryOffs.lsb - currentLBA;
-					currentLBA += pregap_sectors;
-				}
-				currentLBA += GetSizeInSectors(dafile->entry.entrySize.lsb);
-				tinyxml2::XMLElement *newtrack = xmldoc.NewElement(xml::elem::TRACK);
-				newtrack->SetAttribute(xml::attrib::TRACK_TYPE, "audio");
-				if (!dafile->trackid.empty())
-				{
-					newtrack->SetAttribute(xml::attrib::TRACK_ID, dafile->trackid.c_str());
-				}
-				newtrack->SetAttribute(xml::attrib::TRACK_SOURCE, dafile->virtualPath.generic_string().c_str());
-				// only write the pregap element if it's non default
-				if(pregap_sectors != 150)
-				{
-					tinyxml2::XMLElement *pregap = newtrack->InsertNewChildElement(xml::elem::TRACK_PREGAP);
-					pregap->SetAttribute(xml::attrib::PREGAP_DURATION, SectorsToTimecode(pregap_sectors).c_str());
-				}
-
-				modifyProject->InsertAfterChild(addAfter, newtrack);
-				addAfter = newtrack;
-			}
-
-			// Check if there is still an EoF gap
-			if (!param::noWarns && (postGap > 150 || currentLBA < totalLenLBA))
-			{
-				printf( "WARNING: There is still a gap of %u sectors at the end of file system.\n"
-						"\t This could mean that there are missing files or tracks.\n", postGap > 150 ? postGap : totalLenLBA - currentLBA);
-				if (global::cueFile.tracks.empty())
-				{
-					printf("\t Try using a .cue file instead of an ISO image.\n");
-				}
-				else
-				{
-					printf("\t Try using the -pt or/and -f command, helps with obfuscated file systems.\n");
-				}
-			}
-
-			xmldoc.SaveFile(file);
-			fclose(file);
-		}
-		else
-		{
-			printf("\nERROR: Cannot create xml file \"%s\". %s\n", param::xmlFile.string().c_str(), strerror(errno));
-			exit(EXIT_FAILURE);
 		}
 	}
